tools/fmc-fdelay-pulse: report times to the user after parsing
This should in a way fix the missing error checks in parse_time,
without forcing us to rewrite parsing with more stringent checks.

Times are reported before activating rather than after each parse
action, so the user can see what 1mhz or pps turn out to be, as
well as the default values.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>