driver: fix use of TDC user_offset at interrupt time
TDC user offset should be applied to all timestamps. The code, however,
didn't do that properly, and fast_tdc mode was missing it.

Now user_offset is applied where the hardware offset is, like it should
have since time zero.

This should have no effect, as user_offset is an option nobody used
so far, as far as we know.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>