scripts/check-submodule: really check submodules of ../
The should pick other repositories from ../ only if both use and them
are submodules of ../ .

Only checking that ../zio is a directory is not enough, because it
can be a separate clone (old version, for example) or worse an
intermediate directory that doesn't host headers.

A user reported painful build errors for this reason.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>