bld: do not build pre-requirements in real installations this does not work
while this is handy and useful for personal use or a quick start, this
is confusing and problematic for real installations. For this reason I
remove the handy quick-start Makefile rules. People should know how to
build software if they download the sources. Ideally we should deliver
binaries for a set of Linux distributions or DKMS packages for drivers.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>